[Chore] 기록 및 코드 개선
